# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeighbourFit AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface NeighbourhoodAskInput {
  question: string;
  compatibilityScore?: number;
  topPriorities?: string[];
}

export interface NeighbourhoodAskResponse {
  answer: string;
}

export interface CompareOverviewInput {
  /**
   * @minItems 2
   * @maxItems 3
   */
  slugs: string[];
}

export interface CompareOverviewResponse {
  overview: string;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface User {
  id: number;
  email: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message?: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Neighborhood {
  id: number;
  name: string;
  slug: string;
  city: string;
  identity: string;
  /** @nullable */
  description?: string | null;
  affordabilityScore: number;
  walkabilityScore: number;
  transitScore: number;
  nightlifeScore: number;
  safetyScore: number;
  fitnessScore: number;
  petFriendlinessScore: number;
  /** @nullable */
  medianRentalEstimate?: number | null;
  /** @nullable */
  downtownCommuteEstimateMins?: number | null;
  populationDensityClass: string;
  lifestyleTags: string[];
}

export interface NeighborhoodDetail {
  id: number;
  name: string;
  slug: string;
  city: string;
  identity: string;
  /** @nullable */
  description?: string | null;
  affordabilityScore: number;
  walkabilityScore: number;
  transitScore: number;
  nightlifeScore: number;
  safetyScore: number;
  fitnessScore: number;
  petFriendlinessScore: number;
  /** @nullable */
  medianRentalEstimate?: number | null;
  /** @nullable */
  downtownCommuteEstimateMins?: number | null;
  populationDensityClass: string;
  lifestyleTags: string[];
  /** @nullable */
  lastReviewedDate?: string | null;
}

export interface NeighborhoodAdmin {
  id: number;
  name: string;
  slug: string;
  city: string;
  identity: string;
  /** @nullable */
  description?: string | null;
  affordabilityScore: number;
  walkabilityScore: number;
  transitScore: number;
  nightlifeScore: number;
  safetyScore: number;
  fitnessScore: number;
  petFriendlinessScore: number;
  /** @nullable */
  affordabilityScoreNote?: string | null;
  /** @nullable */
  walkabilityScoreNote?: string | null;
  /** @nullable */
  transitScoreNote?: string | null;
  /** @nullable */
  nightlifeScoreNote?: string | null;
  /** @nullable */
  safetyScoreNote?: string | null;
  /** @nullable */
  fitnessScoreNote?: string | null;
  /** @nullable */
  petFriendlinessScoreNote?: string | null;
  /** @nullable */
  medianRentalEstimate?: number | null;
  /** @nullable */
  downtownCommuteEstimateMins?: number | null;
  populationDensityClass: string;
  lifestyleTags: string[];
  /** @nullable */
  lastReviewedDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface NeighborhoodUpdate {
  name?: string;
  identity?: string;
  description?: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  affordabilityScore?: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  walkabilityScore?: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  transitScore?: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  nightlifeScore?: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  safetyScore?: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  fitnessScore?: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  petFriendlinessScore?: number;
  affordabilityScoreNote?: string;
  walkabilityScoreNote?: string;
  transitScoreNote?: string;
  nightlifeScoreNote?: string;
  safetyScoreNote?: string;
  fitnessScoreNote?: string;
  petFriendlinessScoreNote?: string;
  medianRentalEstimate?: number;
  downtownCommuteEstimateMins?: number;
  populationDensityClass?: string;
  lifestyleTags?: string[];
}

export interface UserWeights {
  /**
   * @minimum 0
   * @maximum 100
   */
  affordability: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  walkability: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  transit: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  nightlife: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  safety: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  fitness: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  petFriendliness: number;
}

export interface QuestionnaireInput {
  /** @minimum 0 */
  budget: number;
  weights: UserWeights;
  /** @nullable */
  workplaceNeighborhood?: string | null;
  usedDefaultWeights?: boolean;
}

export interface DimensionBreakdown {
  dimension: string;
  score: number;
  weight: number;
  contribution: number;
}

export interface NeighborhoodMatch {
  neighborhood: Neighborhood;
  compatibilityScore: number;
  fitLabel: string;
  /** @nullable */
  aiSummary?: string | null;
  aiSummaryError: boolean;
  affordabilityWarning: boolean;
  dimensionBreakdown: DimensionBreakdown[];
  tradeoffExplanation: string;
}

export interface RecommendationResult {
  /** @nullable */
  sessionId?: number | null;
  matches: NeighborhoodMatch[];
  usedDefaultWeights: boolean;
  createdAt: string;
}

export interface RecommendationSession {
  id: number;
  budget: number;
  usedDefaultWeights: boolean;
  resultsCount: number;
  createdAt: string;
}

export interface CompareInput {
  /**
   * @minItems 2
   * @maxItems 3
   */
  slugs: string[];
}

export interface NeighborhoodComparison {
  neighborhood: NeighborhoodDetail;
  strengths: string[];
  tradeoffs: string[];
}

export interface Analytics {
  questionnaireStarts: number;
  questionnaireCompletions: number;
  recommendationSessions: number;
  favoriteSaves: number;
  registeredUsers: number;
  guestSessions: number;
  adminNeighborhoodUpdates: number;
  aiSummariesGenerated: number;
  aiCacheHits: number;
  aiCacheMisses: number;
}