# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NeighbourFit AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /neighborhoods/{slug}:
    get:
      operationId: getNeighborhood
      tags: [neighborhoods]
      summary: Get neighborhood by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Neighborhood detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeighborhoodDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /neighborhoods/{slug}/ask:
    post:
      operationId: askNeighbourhood
      tags: [neighborhoods]
      summary: Ask a follow-up question about a neighbourhood
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NeighbourhoodAskInput"
      responses:
        "200":
          description: AI answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeighbourhoodAskResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Recommendations
...
  /neighborhoods/compare-summary:
    post:
      operationId: compareNeighbourhoodsSummary
      tags: [comparison]
      summary: Generate AI comparison overview for 2-3 neighbourhoods
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompareOverviewInput"
      responses:
        "200":
          description: AI comparison overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompareOverviewResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Admin
...
components:
  schemas:
    NeighbourhoodAskInput:
      type: object
      properties:
        question:
          type: string
        compatibilityScore:
          type: number
        topPriorities:
          type: array
          items:
            type: string
      required:
        - question
    NeighbourhoodAskResponse:
      type: object
      properties:
        answer:
          type: string
      required:
        - answer
    CompareOverviewInput:
      type: object
      required:
        - slugs
      properties:
        slugs:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 3
    CompareOverviewResponse:
      type: object
      required:
        - overview
      properties:
        overview:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    NeighborhoodDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        city:
          type: string
        identity:
          type: string
        description:
          type: ["string", "null"]
        affordabilityScore:
          type: integer
        walkabilityScore:
          type: integer
        transitScore:
          type: integer
        nightlifeScore:
          type: integer
        safetyScore:
          type: integer
        fitnessScore:
          type: integer
        petFriendlinessScore:
          type: integer
        medianRentalEstimate:
          type: ["integer", "null"]
        downtownCommuteEstimateMins:
          type: ["integer", "null"]
        populationDensityClass:
          type: string
        lifestyleTags:
          type: array
          items:
            type: string
        lastReviewedDate:
          type: ["string", "null"]
      required:
        - id
        - name
        - slug
        - city
        - identity
        - affordabilityScore
        - walkabilityScore
        - transitScore
        - nightlifeScore
        - safetyScore
        - fitnessScore
        - petFriendlinessScore
        - populationDensityClass
        - lifestyleTags
...