# lib/api-zod/src/generated/types/compareOverviewInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeighbourFit AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompareOverviewInput {
  /**
   * @minItems 2
   * @maxItems 3
   */
  slugs: string[];
}