# lib/api-zod/src/generated/types/compareOverviewResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeighbourFit AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompareOverviewResponse {
  overview: string;
}